+++ packages/chrome/src/elements/header/HeaderItemWrapper.tsx
@@ -5,17 +5,9 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import React, { HTMLAttributes } from 'react';
+import React from 'react';
+import { IHeaderItemWrapperProps } from '../../types';
 import { StyledHeaderItemWrapper } from '../../styled';
-
-export interface IHeaderItemWrapperProps extends HTMLAttributes<HTMLDivElement> {
-  /** Maximizes the width of a flex item in the header */
-  maxX?: boolean;
-  /** Maximizes the height of the item (i.e. contains a search input) */
-  maxY?: boolean;
-  /** Rounds the border radius of the item */
-  isRound?: boolean;
-}
 
 /**
  * @extends HTMLAttributes<HTMLDivElement>
