+++ packages/chrome/src/elements/subnav/SubNavItemText.tsx
@@ -5,17 +5,10 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import React, { HTMLAttributes } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
+import { ISubNavItemTextProps } from '../../types';
 import { StyledSubNavItemText } from '../../styled';
-
-export interface ISubNavItemTextProps extends HTMLAttributes<HTMLSpanElement> {
-  /**
-   * Wraps overflow item text instead of truncating long strings with an ellipsis.
-   * Use when `max-width` styling is applied to the subnav container.
-   **/
-  isWrapped?: boolean;
-}
 
 /**
  * @extends HTMLAttributes<HTMLSpanElement>
