+++ packages/chrome/src/index.ts
@@ -6,40 +6,45 @@
  */
 
 export { Chrome } from './elements/Chrome';
-export type { IChromeProps } from './elements/Chrome';
 export { SkipNav } from './elements/SkipNav';
-export type { ISkipNavProps } from './elements/SkipNav';
 export { Body } from './elements/body/Body';
-export type { IBodyProps } from './elements/body/Body';
 export { Content } from './elements/body/Content';
 export { Main } from './elements/body/Main';
 export { Sidebar } from './elements/body/Sidebar';
 export { Header } from './elements/header/Header';
-export type { IHeaderProps } from './elements/header/Header';
 export { HeaderItem } from './elements/header/HeaderItem';
-export type { IHeaderItemProps } from './elements/header/HeaderItem';
 export { HeaderItemIcon } from './elements/header/HeaderItemIcon';
 export { HeaderItemText } from './elements/header/HeaderItemText';
-export type { IHeaderItemTextProps } from './elements/header/HeaderItemText';
 export { HeaderItemWrapper } from './elements/header/HeaderItemWrapper';
-export type { IHeaderItemWrapperProps } from './elements/header/HeaderItemWrapper';
 export { Footer } from './elements/footer/Footer';
 export { FooterItem } from './elements/footer/FooterItem';
 export { Nav } from './elements/nav/Nav';
-export type { INavProps } from './elements/nav/Nav';
 export { NavItem } from './elements/nav/NavItem';
-export type { INavItemProps } from './elements/nav/NavItem';
 export { NavItemIcon } from './elements/nav/NavItemIcon';
 export { NavItemText } from './elements/nav/NavItemText';
-export type { INavItemTextProps } from './elements/nav/NavItemText';
 export { SubNav } from './elements/subnav/SubNav';
 export { SubNavItem } from './elements/subnav/SubNavItem';
-export type { ISubNavItemProps } from './elements/subnav/SubNavItem';
 export { SubNavItemText } from './elements/subnav/SubNavItemText';
-export type { ISubNavItemTextProps } from './elements/subnav/SubNavItemText';
 export { CollapsibleSubNavItem } from './elements/subnav/CollapsibleSubNavItem';
-export type { ICollapsibleSubNavItemProps } from './elements/subnav/CollapsibleSubNavItem';
 export { Sheet } from './elements/sheet/Sheet';
-export type { ISheetProps } from './elements/sheet/Sheet';
-export type { ISheetFooterProps } from './elements/sheet/components/Footer';
-export { PRODUCT, PRODUCTS } from './utils/types';
+
+export { PRODUCT as PRODUCTS } from './types';
+
+export type {
+  IChromeProps,
+  ISkipNavProps,
+  IBodyProps,
+  IHeaderProps,
+  IHeaderItemProps,
+  IHeaderItemTextProps,
+  IHeaderItemWrapperProps,
+  INavProps,
+  INavItemProps,
+  INavItemTextProps,
+  ISubNavItemProps,
+  ICollapsibleSubNavItemProps,
+  ISheetProps,
+  ISheetFooterProps,
+  /** @deprecated can be accessed via IHeaderItemProps['product'] */
+  Product as PRODUCT
+} from './types';
