--- packages/chrome/src/utils/types.ts
@@ -1,18 +1,0 @@
-/**
- * Copyright Zendesk, Inc.
- *
- * Use of this source code is governed under the Apache License, Version 2.0
- * found at http://www.apache.org/licenses/LICENSE-2.0.
- */
-
-export type PRODUCT = 'chat' | 'connect' | 'explore' | 'guide' | 'message' | 'support' | 'talk';
-
-export const PRODUCTS: PRODUCT[] = [
-  'chat',
-  'connect',
-  'explore',
-  'guide',
-  'message',
-  'support',
-  'talk'
-];
